add limited support for overlapped atoms
when inserting an atom on top of another atom, set a flag on the
destination atom and put the pending atom in an array.  when the
destination atom is removed, check the flag and pull the pending atom
out of the array.

thanks to grimmy for writing the initial version of this change and
inspiring the approach used.